requirements.txt
langchain==1.2.10
langchain-core==1.2.13
langchain-anthropic==1.3.3
langchain-openai==1.1.9
langgraph==1.0.8
python-dotenv==1.1.1
